backend/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9

# Face Recognition
deepface==0.0.93
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
scikit-learn==1.4.2
pandas==2.2.2

# Logging
loguru==0.7.2

# HTTP client (for DeepFace model download)
httpx==0.27.0
passlib
bcrypt
python-jose[cryptography]
python-multipart
email-validator
pydantic[email]
